Fix indirect draw test for MSVC
We need a negative number in a buffer of uint32_t, to test negative
baseVertex. MSVC requires an explicit cast here.

Bug: dawn:172

+++ src/tests/end2end/DrawIndexedIndirectTests.cpp
@@ -137,10 +137,10 @@ TEST_P(DrawIndexedIndirectTest, BaseVertex) {
 
     // Test negative baseVertex
     // Test a draw with only the first 3 indices of the first quad (bottom left triangle)
-    Test({3, 1, 0, -4, 0}, 6 * sizeof(uint32_t), 0, filled, notFilled);
+    Test({3, 1, 0, static_cast<uint32_t>(-4), 0}, 6 * sizeof(uint32_t), 0, filled, notFilled);
 
     // Test a draw with only the last 3 indices of the first quad (top right triangle)
-    Test({3, 1, 3, -4, 0}, 6 * sizeof(uint32_t), 0, notFilled, filled);
+    Test({3, 1, 3, static_cast<uint32_t>(-4), 0}, 6 * sizeof(uint32_t), 0, notFilled, filled);
 }
 
 TEST_P(DrawIndexedIndirectTest, IndirectOffset) {
